Odroid N2 : Cleaned up fix for ethernet controller issues
Sometimes the controller becomes unavailable after a software reboot,
or it is not possible to up & down Ethernet as pleased.

This commit cleans up the existing fix by using the proper value for the reset timing.

This has been fixed in the 5.10 kernel from kernel.org :
https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?id=a83bf1290710b82c420f9a33808d8d69f90f0eb3
https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/log/?qt=grep&q=RTL8211F&h=linux-5.10.y

The fix will be removed when the Debian kernel in the buster-backports upgrades to a 5.10.